change to get_video_paths for clarity
